+++ src/components/Navbar/index.tsx
@@ -7,6 +7,7 @@
 
 import { CFC } from "@typings/react";
 import NavbarWrapper from "./Wrapper";
+import NavbarBrand from "./Brand";
 
 const Navbar: CFC<HTMLDivElement> = ({ children, ...rest }) => {
   return (
@@ -21,4 +22,5 @@ const Navbar: CFC<HTMLDivElement> = ({ children, ...rest }) => {
 
 export default Object.assign(Navbar, {
   Wrapper: NavbarWrapper,
+  Brand: NavbarBrand,
 });